requirements.txt
flask==2.3.3
openai==1.3.0
python-dotenv==1.0.0
flask-cors==4.0.0 